requirements.txt
numpy
pandas
pillow
keras
streamlit
